refactor(components,fields): one fullscreen long-text editor, hoisted to components (#3398)
The "expand to a full-height dialog" interaction had two independent
implementations: `FullscreenTextarea` inside the form renderer's built-in
`textarea` branch (@object-ui/components) and `FullscreenFieldEditor` in
@object-ui/fields for the registered TextAreaField / RichTextField widgets.
One form-level promise (ObjectFormSchema.mobile.fullscreenLongText) is
honoured on two render paths, and each path grew its own answer.

They drifted in both directions — #3400 (readonly editable through the
built-in dialog), #3402 (the same write-back hole for disabled on the
registered path), #3393 and #3272 each landing on one side first.

Per the maintainer ruling of 2026-08-10, the shared primitive is hoisted to
the dependency-legal side. Measured import graph: fields depends on
components; components declares no dependency on fields in dependencies or
peerDependencies. So components now exports `FullscreenEditor` and
`FullscreenFieldEditor` becomes a thin wrapper over it.

The primitive DEFINES readOnly/disabled rather than inheriting them by
accident: readOnly renders no affordance at all, disabled leaves an inert
one, and both are gated at open, in the injected editor, on Done, and at
onCommit — because disabled also carries isSubmitting and can flip while the
dialog is open.

Copy is unchanged: the same form.fullscreen.* / common.cancel keys through
createSafeTranslation with byte-identical English defaults, so no locale pack
changes and provider-less hosts render what they did. The now-unread
form.fullscreen.* defaults are dropped from useFieldTranslation.

`toggleClassName` is not carried into the new primitive — zero producers
repo-wide, and FullscreenFieldEditor is not exported from the fields barrel.
